Add context manager, extract repeated code block to new function #320
Reference: https://github.com/nexB/purldb/issues/320
