+++ purldb-toolkit/tests/test_purlcli.py
@@ -10,8 +10,12 @@
 import json
 import os
 from collections import OrderedDict
+from contextlib import redirect_stdout
+from io import StringIO
+from unittest import mock
 
 import pytest
+import requests
 from click.testing import CliRunner
 from commoncode.testcase import FileDrivenTesting
 from purldb_toolkit import cli_test_utils, purlcli
@@ -19,202 +23,8 @@ from purldb_toolkit import cli_test_utils, purlcli
 test_env = FileDrivenTesting()
 test_env.test_data_dir = os.path.join(os.path.dirname(__file__), "data")
 
-pytestmark = pytest.mark.live_fetch()
-
 
 class TestPURLCLI_metadata(object):
-    def test_metadata_cli(self):
-        """
-        Test the `metadata` command with actual and expected JSON output files.
-
-        Note that we can't simply compare the actual and expected JSON files
-        because the `--output` values (paths) differ due to the use of
-        temporary files, and therefore we test a list of relevant key-value pairs.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_metadata_output.json"
-        )
-        actual_result_file = test_env.get_temp_file("actual_metadata_output.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0?os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:gem/rails",
-            "--purl",
-            "pkg:rubygems/rails",
-            "--output",
-            actual_result_file,
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.get_metadata, options, catch_exceptions=False)
-        assert result.exit_code == 0
-
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-        cli_test_utils.streamline_headers(output_data["headers"])
-        streamline_metadata_packages(output_data["packages"])
-
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-        cli_test_utils.streamline_headers(expected_data["headers"])
-        streamline_metadata_packages(expected_data["packages"])
-
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
-        ]
-
-        for output, expected in result_objects:
-            assert output == expected
-
-        # NOTE: To avoid errors from the addition of new versions, we exclude
-        # "packages" from the result_objects list above and handle here.
-        # All other live-fetch tests are handled in a similar manner.
-
-        expected_metadata = []
-        for pkg in expected_data["packages"]:
-            expected_metadata.append(pkg)
-
-        actual_metadata = []
-        for pkg in output_data["packages"]:
-            actual_metadata.append(pkg)
-
-        for exp in expected_metadata:
-            for actual in actual_metadata:
-                if exp["purl"] == actual["purl"]:
-                    for exp_meta in exp["metadata"]:
-                        assert exp_meta in actual["metadata"]
-
-    def test_metadata_cli_unique(self):
-        """
-        Test the `metadata` command with actual and expected JSON output files
-        with the `--unique` flag included in the command.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_metadata_output_unique.json"
-        )
-        actual_result_file = test_env.get_temp_file("actual_metadata_output.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0?os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:gem/rails",
-            "--purl",
-            "pkg:rubygems/rails",
-            "--output",
-            actual_result_file,
-            "--unique",
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.get_metadata, options, catch_exceptions=False)
-        assert result.exit_code == 0
-
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-        cli_test_utils.streamline_headers(output_data["headers"])
-        streamline_metadata_packages(output_data["packages"])
-
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-        cli_test_utils.streamline_headers(expected_data["headers"])
-        streamline_metadata_packages(expected_data["packages"])
-
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--unique"],
-                expected_data["headers"][0]["options"]["--unique"],
-            ),
-        ]
-
-        for output, expected in result_objects:
-            assert output == expected
-
-        expected_metadata = []
-        for pkg in expected_data["packages"]:
-            expected_metadata.append(pkg)
-
-        actual_metadata = []
-        for pkg in output_data["packages"]:
-            actual_metadata.append(pkg)
-
-        for exp in expected_metadata:
-            for actual in actual_metadata:
-                if exp["purl"] == actual["purl"]:
-                    for exp_meta in exp["metadata"]:
-                        assert exp_meta in actual["metadata"]
-
     def test_metadata_cli_duplicate_input_sources(self):
         """
         Test the `metadata` command with both `--purl` and `--file` inputs.
@@ -245,479 +55,463 @@ class TestPURLCLI_metadata(object):
         assert "Use either purls or file." in result.output
         assert result.exit_code == 2
 
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                ["pkg:pypi/fetchcode"],
+    @mock.patch("purldb_toolkit.purlcli.collect_metadata")
+    @mock.patch("purldb_toolkit.purlcli.check_metadata_purl")
+    def test_metadata_details(self, mock_check_metadata_purl, mock_collect_metadata):
+
+        mock_collect_metadata.return_value = [
+            OrderedDict(
+                [
+                    ("purl", "pkg:pypi/fetchcode"),
+                    ("type", "pypi"),
+                    ("namespace", None),
+                    ("name", "fetchcode"),
+                    ("version", None),
+                    ("qualifiers", OrderedDict()),
+                    ("subpath", None),
+                    ("primary_language", None),
+                    ("description", None),
+                    ("release_date", None),
+                    ("parties", []),
+                    ("keywords", []),
+                    ("homepage_url", "https://github.com/nexB/fetchcode"),
+                    ("download_url", None),
+                    ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                    ("size", None),
+                    ("sha1", None),
+                    ("md5", None),
+                    ("sha256", None),
+                    ("sha512", None),
+                    ("bug_tracking_url", None),
+                    ("code_view_url", None),
+                    ("vcs_url", None),
+                    ("copyright", None),
+                    ("license_expression", None),
+                    ("declared_license", "Apache-2.0"),
+                    ("notice_text", None),
+                    ("root_path", None),
+                    ("dependencies", []),
+                    ("contains_source_code", None),
+                    ("source_packages", []),
+                    ("repository_homepage_url", None),
+                    ("repository_download_url", None),
+                    ("api_data_url", None),
+                ]
+            ),
+            OrderedDict(
+                [
+                    ("purl", "pkg:pypi/fetchcode@0.1.0"),
+                    ("type", "pypi"),
+                    ("namespace", None),
+                    ("name", "fetchcode"),
+                    ("version", "0.1.0"),
+                    ("qualifiers", OrderedDict()),
+                    ("subpath", None),
+                    ("primary_language", None),
+                    ("description", None),
+                    ("release_date", None),
+                    ("parties", []),
+                    ("keywords", []),
+                    ("homepage_url", "https://github.com/nexB/fetchcode"),
+                    (
+                        "download_url",
+                        "https://files.pythonhosted.org/packages/19/a0/c90e5ba4d71ea1a1a89784f6d839ffb0dbf32d270cba04d5602188cb3713/fetchcode-0.1.0-py3-none-any.whl",
+                    ),
+                    ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                    ("size", None),
+                    ("sha1", None),
+                    ("md5", None),
+                    ("sha256", None),
+                    ("sha512", None),
+                    ("bug_tracking_url", None),
+                    ("code_view_url", None),
+                    ("vcs_url", None),
+                    ("copyright", None),
+                    ("license_expression", None),
+                    ("declared_license", "Apache-2.0"),
+                    ("notice_text", None),
+                    ("root_path", None),
+                    ("dependencies", []),
+                    ("contains_source_code", None),
+                    ("source_packages", []),
+                    ("repository_homepage_url", None),
+                    ("repository_download_url", None),
+                    ("api_data_url", None),
+                ]
+            ),
+            OrderedDict(
+                [
+                    ("purl", "pkg:pypi/fetchcode@0.2.0"),
+                    ("type", "pypi"),
+                    ("namespace", None),
+                    ("name", "fetchcode"),
+                    ("version", "0.2.0"),
+                    ("qualifiers", OrderedDict()),
+                    ("subpath", None),
+                    ("primary_language", None),
+                    ("description", None),
+                    ("release_date", None),
+                    ("parties", []),
+                    ("keywords", []),
+                    ("homepage_url", "https://github.com/nexB/fetchcode"),
+                    (
+                        "download_url",
+                        "https://files.pythonhosted.org/packages/d7/e9/96e9302e84e326b3c10a40c1723f21f4db96b557a17c6871e7a4c6336906/fetchcode-0.2.0-py3-none-any.whl",
+                    ),
+                    ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                    ("size", None),
+                    ("sha1", None),
+                    ("md5", None),
+                    ("sha256", None),
+                    ("sha512", None),
+                    ("bug_tracking_url", None),
+                    ("code_view_url", None),
+                    ("vcs_url", None),
+                    ("copyright", None),
+                    ("license_expression", None),
+                    ("declared_license", "Apache-2.0"),
+                    ("notice_text", None),
+                    ("root_path", None),
+                    ("dependencies", []),
+                    ("contains_source_code", None),
+                    ("source_packages", []),
+                    ("repository_homepage_url", None),
+                    ("repository_download_url", None),
+                    ("api_data_url", None),
+                ]
+            ),
+            OrderedDict(
+                [
+                    ("purl", "pkg:pypi/fetchcode@0.3.0"),
+                    ("type", "pypi"),
+                    ("namespace", None),
+                    ("name", "fetchcode"),
+                    ("version", "0.3.0"),
+                    ("qualifiers", OrderedDict()),
+                    ("subpath", None),
+                    ("primary_language", None),
+                    ("description", None),
+                    ("release_date", None),
+                    ("parties", []),
+                    ("keywords", []),
+                    ("homepage_url", "https://github.com/nexB/fetchcode"),
+                    (
+                        "download_url",
+                        "https://files.pythonhosted.org/packages/8d/fb/e45da0abf63504c3f88ad02537dc9dc64ea5206b09ce29cfb8191420d678/fetchcode-0.3.0-py3-none-any.whl",
+                    ),
+                    ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                    ("size", None),
+                    ("sha1", None),
+                    ("md5", None),
+                    ("sha256", None),
+                    ("sha512", None),
+                    ("bug_tracking_url", None),
+                    ("code_view_url", None),
+                    ("vcs_url", None),
+                    ("copyright", None),
+                    ("license_expression", None),
+                    ("declared_license", "Apache-2.0"),
+                    ("notice_text", None),
+                    ("root_path", None),
+                    ("dependencies", []),
+                    ("contains_source_code", None),
+                    ("source_packages", []),
+                    ("repository_homepage_url", None),
+                    ("repository_download_url", None),
+                    ("api_data_url", None),
+                ]
+            ),
+        ]
+
+        mock_check_metadata_purl.return_value = None
+
+        expected_data = {
+            "headers": [
                 {
-                    "headers": [
-                        {
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "options": {
-                                "command": "metadata",
-                                "--purl": ["pkg:pypi/fetchcode"],
-                                "--file": None,
-                                "--output": "",
-                            },
-                            "purls": ["pkg:pypi/fetchcode"],
-                            "errors": [],
-                            "warnings": [],
-                        }
-                    ],
-                    "packages": [
-                        {
-                            "purl": "pkg:pypi/fetchcode",
-                            "metadata": [
-                                OrderedDict(
-                                    [
-                                        ("purl", "pkg:pypi/fetchcode"),
-                                        ("type", "pypi"),
-                                        ("namespace", None),
-                                        ("name", "fetchcode"),
-                                        ("version", None),
-                                        ("qualifiers", OrderedDict()),
-                                        ("subpath", None),
-                                        ("primary_language", None),
-                                        ("description", None),
-                                        ("release_date", None),
-                                        ("parties", []),
-                                        ("keywords", []),
-                                        (
-                                            "homepage_url",
-                                            "https://github.com/nexB/fetchcode",
-                                        ),
-                                        ("download_url", None),
-                                        (
-                                            "api_url",
-                                            "https://pypi.org/pypi/fetchcode/json",
-                                        ),
-                                        ("size", None),
-                                        ("sha1", None),
-                                        ("md5", None),
-                                        ("sha256", None),
-                                        ("sha512", None),
-                                        ("bug_tracking_url", None),
-                                        ("code_view_url", None),
-                                        ("vcs_url", None),
-                                        ("copyright", None),
-                                        ("license_expression", None),
-                                        ("declared_license", "Apache-2.0"),
-                                        ("notice_text", None),
-                                        ("root_path", None),
-                                        ("dependencies", []),
-                                        ("contains_source_code", None),
-                                        ("source_packages", []),
-                                        ("repository_homepage_url", None),
-                                        ("repository_download_url", None),
-                                        ("api_data_url", None),
-                                    ]
+                    "tool_name": "purlcli",
+                    "tool_version": "0.1.0",
+                    "options": {
+                        "command": "metadata",
+                        "--purl": ["pkg:pypi/fetchcode"],
+                        "--file": None,
+                        "--output": "",
+                    },
+                    "purls": ["pkg:pypi/fetchcode"],
+                    "errors": [],
+                    "warnings": [],
+                }
+            ],
+            "packages": [
+                {
+                    "purl": "pkg:pypi/fetchcode",
+                    "metadata": [
+                        OrderedDict(
+                            [
+                                ("purl", "pkg:pypi/fetchcode"),
+                                ("type", "pypi"),
+                                ("namespace", None),
+                                ("name", "fetchcode"),
+                                ("version", None),
+                                ("qualifiers", OrderedDict()),
+                                ("subpath", None),
+                                ("primary_language", None),
+                                ("description", None),
+                                ("release_date", None),
+                                ("parties", []),
+                                ("keywords", []),
+                                ("homepage_url", "https://github.com/nexB/fetchcode"),
+                                ("download_url", None),
+                                ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                                ("size", None),
+                                ("sha1", None),
+                                ("md5", None),
+                                ("sha256", None),
+                                ("sha512", None),
+                                ("bug_tracking_url", None),
+                                ("code_view_url", None),
+                                ("vcs_url", None),
+                                ("copyright", None),
+                                ("license_expression", None),
+                                ("declared_license", "Apache-2.0"),
+                                ("notice_text", None),
+                                ("root_path", None),
+                                ("dependencies", []),
+                                ("contains_source_code", None),
+                                ("source_packages", []),
+                                ("repository_homepage_url", None),
+                                ("repository_download_url", None),
+                                ("api_data_url", None),
+                            ]
+                        ),
+                        OrderedDict(
+                            [
+                                ("purl", "pkg:pypi/fetchcode@0.1.0"),
+                                ("type", "pypi"),
+                                ("namespace", None),
+                                ("name", "fetchcode"),
+                                ("version", "0.1.0"),
+                                ("qualifiers", OrderedDict()),
+                                ("subpath", None),
+                                ("primary_language", None),
+                                ("description", None),
+                                ("release_date", None),
+                                ("parties", []),
+                                ("keywords", []),
+                                ("homepage_url", "https://github.com/nexB/fetchcode"),
+                                (
+                                    "download_url",
+                                    "https://files.pythonhosted.org/packages/19/a0/c90e5ba4d71ea1a1a89784f6d839ffb0dbf32d270cba04d5602188cb3713/fetchcode-0.1.0-py3-none-any.whl",
                                 ),
-                                OrderedDict(
-                                    [
-                                        ("purl", "pkg:pypi/fetchcode@0.1.0"),
-                                        ("type", "pypi"),
-                                        ("namespace", None),
-                                        ("name", "fetchcode"),
-                                        ("version", "0.1.0"),
-                                        ("qualifiers", OrderedDict()),
-                                        ("subpath", None),
-                                        ("primary_language", None),
-                                        ("description", None),
-                                        ("release_date", None),
-                                        ("parties", []),
-                                        ("keywords", []),
-                                        (
-                                            "homepage_url",
-                                            "https://github.com/nexB/fetchcode",
-                                        ),
-                                        (
-                                            "download_url",
-                                            "https://files.pythonhosted.org/packages/19/a0/c90e5ba4d71ea1a1a89784f6d839ffb0dbf32d270cba04d5602188cb3713/fetchcode-0.1.0-py3-none-any.whl",
-                                        ),
-                                        (
-                                            "api_url",
-                                            "https://pypi.org/pypi/fetchcode/json",
-                                        ),
-                                        ("size", None),
-                                        ("sha1", None),
-                                        ("md5", None),
-                                        ("sha256", None),
-                                        ("sha512", None),
-                                        ("bug_tracking_url", None),
-                                        ("code_view_url", None),
-                                        ("vcs_url", None),
-                                        ("copyright", None),
-                                        ("license_expression", None),
-                                        ("declared_license", "Apache-2.0"),
-                                        ("notice_text", None),
-                                        ("root_path", None),
-                                        ("dependencies", []),
-                                        ("contains_source_code", None),
-                                        ("source_packages", []),
-                                        ("repository_homepage_url", None),
-                                        ("repository_download_url", None),
-                                        ("api_data_url", None),
-                                    ]
+                                ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                                ("size", None),
+                                ("sha1", None),
+                                ("md5", None),
+                                ("sha256", None),
+                                ("sha512", None),
+                                ("bug_tracking_url", None),
+                                ("code_view_url", None),
+                                ("vcs_url", None),
+                                ("copyright", None),
+                                ("license_expression", None),
+                                ("declared_license", "Apache-2.0"),
+                                ("notice_text", None),
+                                ("root_path", None),
+                                ("dependencies", []),
+                                ("contains_source_code", None),
+                                ("source_packages", []),
+                                ("repository_homepage_url", None),
+                                ("repository_download_url", None),
+                                ("api_data_url", None),
+                            ]
+                        ),
+                        OrderedDict(
+                            [
+                                ("purl", "pkg:pypi/fetchcode@0.2.0"),
+                                ("type", "pypi"),
+                                ("namespace", None),
+                                ("name", "fetchcode"),
+                                ("version", "0.2.0"),
+                                ("qualifiers", OrderedDict()),
+                                ("subpath", None),
+                                ("primary_language", None),
+                                ("description", None),
+                                ("release_date", None),
+                                ("parties", []),
+                                ("keywords", []),
+                                ("homepage_url", "https://github.com/nexB/fetchcode"),
+                                (
+                                    "download_url",
+                                    "https://files.pythonhosted.org/packages/d7/e9/96e9302e84e326b3c10a40c1723f21f4db96b557a17c6871e7a4c6336906/fetchcode-0.2.0-py3-none-any.whl",
                                 ),
-                                OrderedDict(
-                                    [
-                                        ("purl", "pkg:pypi/fetchcode@0.2.0"),
-                                        ("type", "pypi"),
-                                        ("namespace", None),
-                                        ("name", "fetchcode"),
-                                        ("version", "0.2.0"),
-                                        ("qualifiers", OrderedDict()),
-                                        ("subpath", None),
-                                        ("primary_language", None),
-                                        ("description", None),
-                                        ("release_date", None),
-                                        ("parties", []),
-                                        ("keywords", []),
-                                        (
-                                            "homepage_url",
-                                            "https://github.com/nexB/fetchcode",
-                                        ),
-                                        (
-                                            "download_url",
-                                            "https://files.pythonhosted.org/packages/d7/e9/96e9302e84e326b3c10a40c1723f21f4db96b557a17c6871e7a4c6336906/fetchcode-0.2.0-py3-none-any.whl",
-                                        ),
-                                        (
-                                            "api_url",
-                                            "https://pypi.org/pypi/fetchcode/json",
-                                        ),
-                                        ("size", None),
-                                        ("sha1", None),
-                                        ("md5", None),
-                                        ("sha256", None),
-                                        ("sha512", None),
-                                        ("bug_tracking_url", None),
-                                        ("code_view_url", None),
-                                        ("vcs_url", None),
-                                        ("copyright", None),
-                                        ("license_expression", None),
-                                        ("declared_license", "Apache-2.0"),
-                                        ("notice_text", None),
-                                        ("root_path", None),
-                                        ("dependencies", []),
-                                        ("contains_source_code", None),
-                                        ("source_packages", []),
-                                        ("repository_homepage_url", None),
-                                        ("repository_download_url", None),
-                                        ("api_data_url", None),
-                                    ]
+                                ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                                ("size", None),
+                                ("sha1", None),
+                                ("md5", None),
+                                ("sha256", None),
+                                ("sha512", None),
+                                ("bug_tracking_url", None),
+                                ("code_view_url", None),
+                                ("vcs_url", None),
+                                ("copyright", None),
+                                ("license_expression", None),
+                                ("declared_license", "Apache-2.0"),
+                                ("notice_text", None),
+                                ("root_path", None),
+                                ("dependencies", []),
+                                ("contains_source_code", None),
+                                ("source_packages", []),
+                                ("repository_homepage_url", None),
+                                ("repository_download_url", None),
+                                ("api_data_url", None),
+                            ]
+                        ),
+                        OrderedDict(
+                            [
+                                ("purl", "pkg:pypi/fetchcode@0.3.0"),
+                                ("type", "pypi"),
+                                ("namespace", None),
+                                ("name", "fetchcode"),
+                                ("version", "0.3.0"),
+                                ("qualifiers", OrderedDict()),
+                                ("subpath", None),
+                                ("primary_language", None),
+                                ("description", None),
+                                ("release_date", None),
+                                ("parties", []),
+                                ("keywords", []),
+                                ("homepage_url", "https://github.com/nexB/fetchcode"),
+                                (
+                                    "download_url",
+                                    "https://files.pythonhosted.org/packages/8d/fb/e45da0abf63504c3f88ad02537dc9dc64ea5206b09ce29cfb8191420d678/fetchcode-0.3.0-py3-none-any.whl",
                                 ),
-                                OrderedDict(
-                                    [
-                                        ("purl", "pkg:pypi/fetchcode@0.3.0"),
-                                        ("type", "pypi"),
-                                        ("namespace", None),
-                                        ("name", "fetchcode"),
-                                        ("version", "0.3.0"),
-                                        ("qualifiers", OrderedDict()),
-                                        ("subpath", None),
-                                        ("primary_language", None),
-                                        ("description", None),
-                                        ("release_date", None),
-                                        ("parties", []),
-                                        ("keywords", []),
-                                        (
-                                            "homepage_url",
-                                            "https://github.com/nexB/fetchcode",
-                                        ),
-                                        (
-                                            "download_url",
-                                            "https://files.pythonhosted.org/packages/8d/fb/e45da0abf63504c3f88ad02537dc9dc64ea5206b09ce29cfb8191420d678/fetchcode-0.3.0-py3-none-any.whl",
-                                        ),
-                                        (
-                                            "api_url",
-                                            "https://pypi.org/pypi/fetchcode/json",
-                                        ),
-                                        ("size", None),
-                                        ("sha1", None),
-                                        ("md5", None),
-                                        ("sha256", None),
-                                        ("sha512", None),
-                                        ("bug_tracking_url", None),
-                                        ("code_view_url", None),
-                                        ("vcs_url", None),
-                                        ("copyright", None),
-                                        ("license_expression", None),
-                                        ("declared_license", "Apache-2.0"),
-                                        ("notice_text", None),
-                                        ("root_path", None),
-                                        ("dependencies", []),
-                                        ("contains_source_code", None),
-                                        ("source_packages", []),
-                                        ("repository_homepage_url", None),
-                                        ("repository_download_url", None),
-                                        ("api_data_url", None),
-                                    ]
-                                ),
-                            ],
-                        }
+                                ("api_url", "https://pypi.org/pypi/fetchcode/json"),
+                                ("size", None),
+                                ("sha1", None),
+                                ("md5", None),
+                                ("sha256", None),
+                                ("sha512", None),
+                                ("bug_tracking_url", None),
+                                ("code_view_url", None),
+                                ("vcs_url", None),
+                                ("copyright", None),
+                                ("license_expression", None),
+                                ("declared_license", "Apache-2.0"),
+                                ("notice_text", None),
+                                ("root_path", None),
+                                ("dependencies", []),
+                                ("contains_source_code", None),
+                                ("source_packages", []),
+                                ("repository_homepage_url", None),
+                                ("repository_download_url", None),
+                                ("api_data_url", None),
+                            ]
+                        ),
                     ],
-                },
-            ),
-            (
-                ["pkg:gem/bundler-sass"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:gem/bundler-sass"],
-                                "command": "metadata",
-                            },
-                            "purls": ["pkg:gem/bundler-sass"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": [
-                                "'pkg:gem/bundler-sass' not supported with `metadata` command"
-                            ],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["pkg:rubygems/bundler-sass"],
-                {
-                    "headers": [
-                        {
-                            "tool_name": "purlcli",
-                            "options": {
-                                "command": "metadata",
-                                "--purl": ["pkg:rubygems/bundler-sass"],
-                                "--file": None,
-                                "--output": "",
-                            },
-                            "purls": ["pkg:rubygems/bundler-sass"],
-                            "errors": [],
-                            "warnings": [
-                                "'check_existence' is not supported for "
-                                "'pkg:rubygems/bundler-sass'",
-                            ],
-                        }
-                    ],
-                    "packages": [
-                        {
-                            "purl": "pkg:rubygems/bundler-sass",
-                            "metadata": [
-                                OrderedDict(
-                                    [
-                                        ("purl", "pkg:rubygems/bundler-sass"),
-                                        ("type", "rubygems"),
-                                        ("namespace", None),
-                                        ("name", "bundler-sass"),
-                                        ("version", None),
-                                        ("qualifiers", OrderedDict()),
-                                        ("subpath", None),
-                                        ("primary_language", None),
-                                        ("description", None),
-                                        ("release_date", None),
-                                        ("parties", []),
-                                        ("keywords", []),
-                                        (
-                                            "homepage_url",
-                                            "http://github.com/vogelbek/bundler-sass",
-                                        ),
-                                        (
-                                            "download_url",
-                                            "https://rubygems.org/gems/bundler-sass-0.1.2.gem",
-                                        ),
-                                        (
-                                            "api_url",
-                                            "https://rubygems.org/api/v1/gems/bundler-sass.json",
-                                        ),
-                                        ("size", None),
-                                        ("sha1", None),
-                                        ("md5", None),
-                                        ("sha256", None),
-                                        ("sha512", None),
-                                        ("bug_tracking_url", None),
-                                        ("code_view_url", None),
-                                        ("vcs_url", None),
-                                        ("copyright", None),
-                                        ("license_expression", None),
-                                        ("declared_license", ["MIT"]),
-                                        ("notice_text", None),
-                                        ("root_path", None),
-                                        ("dependencies", []),
-                                        ("contains_source_code", None),
-                                        ("source_packages", []),
-                                        ("repository_homepage_url", None),
-                                        ("repository_download_url", None),
-                                        ("api_data_url", None),
-                                    ]
-                                )
-                            ],
-                        }
-                    ],
-                },
-            ),
-            (
-                ["pkg:nginx/nginx"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:nginx/nginx"],
-                                "command": "metadata",
-                            },
-                            "purls": ["pkg:nginx/nginx"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": [
-                                "'pkg:nginx/nginx' not supported with `metadata` command"
-                            ],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["pkg:pypi/zzzzz"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:pypi/zzzzz"],
-                                "command": "metadata",
-                            },
-                            "purls": ["pkg:pypi/zzzzz"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": [
-                                "'pkg:pypi/zzzzz' does not exist in the upstream repo",
-                            ],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["pkg:pypi/?fetchcode"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:pypi/?fetchcode"],
-                                "command": "metadata",
-                            },
-                            "purls": ["pkg:pypi/?fetchcode"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": ["'pkg:pypi/?fetchcode' not valid"],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["zzzzz"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["zzzzz"],
-                                "command": "metadata",
-                            },
-                            "purls": ["zzzzz"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": ["'zzzzz' not valid"],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-        ],
-    )
-    def test_metadata_details(self, test_input, expected):
-        """
-        Test the `metadata` nested function, `get_metadata_details()`.
-        """
-        purl_metadata = purlcli.get_metadata_details(
-            test_input,
-            output="",
-            file="",
-            command_name="metadata",
-            unique=False,
+                }
+            ],
+        }
+
+        input_purls = ["pkg:pypi/fetchcode"]
+
+        output = ""
+        file = ""
+        command_name = "metadata"
+        unique = False
+
+        purl_metadata_data = purlcli.get_metadata_details(
+            input_purls,
+            output,
+            file,
+            unique,
+            command_name,
         )
-        cli_test_utils.streamline_headers(purl_metadata["headers"])
-        streamline_metadata_packages(purl_metadata["packages"])
 
-        cli_test_utils.streamline_headers(expected["headers"])
-        streamline_metadata_packages(expected["packages"])
+        cli_test_utils.streamline_headers(purl_metadata_data["headers"])
+        cli_test_utils.streamline_headers(expected_data["headers"])
 
-        assert purl_metadata["headers"] == expected["headers"]
+        assert purl_metadata_data == expected_data
 
-        expected_metadata = []
-        for pkg in expected["packages"]:
-            expected_metadata.append(pkg)
+    @mock.patch("purldb_toolkit.purlcli.validate_purl")
+    def test_check_metadata_purl(self, mock_validate_purl):
+        mock_validate_purl.return_value = {
+            "valid": True,
+            "exists": None,
+            "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
+            "purl": "pkg:rubygems/bundler-sass",
+        }
+        input_purl = "pkg:rubygems/bundler-sass"
+        expected = "check_existence_not_supported"
+        purl_metadata = purlcli.check_metadata_purl(input_purl)
 
-        actual_metadata = []
-        for pkg in purl_metadata["packages"]:
-            actual_metadata.append(pkg)
-
-        for exp in expected_metadata:
-            for actual in actual_metadata:
-                if exp["purl"] == actual["purl"]:
-                    for exp_meta in exp["metadata"]:
-                        assert exp_meta in actual["metadata"]
-
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                ["pkg:pypi/fetchcode"],
-                None,
-            ),
-            (
-                ["pkg:gem/bundler-sass"],
-                "valid_but_not_supported",
-            ),
-            (
-                ["pkg:rubygems/bundler-sass"],
-                "check_existence_not_supported",
-            ),
-            (
-                ["pkg:nginx/nginx"],
-                "valid_but_not_supported",
-            ),
-            (
-                ["pkg:pypi/zzzzz"],
-                "not_in_upstream_repo",
-            ),
-            (
-                ["pkg:pypi/?fetchcode"],
-                "not_valid",
-            ),
-            (
-                ["zzzzz"],
-                "not_valid",
-            ),
-        ],
-    )
-    def test_check_metadata_purl(self, test_input, expected):
-        purl_metadata = purlcli.check_metadata_purl(test_input[0])
         assert purl_metadata == expected
+
+    @mock.patch("purldb_toolkit.purlcli.validate_purl")
+    def test_check_metadata_purl_multiple(self, mock_validate_purl):
+        mock_validate_purl.side_effect = [
+            {
+                "valid": True,
+                "exists": True,
+                "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+                "purl": "pkg:pypi/fetchcode",
+            },
+            {
+                "valid": True,
+                "exists": True,
+                "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+                "purl": "pkg:gem/bundler-sass",
+            },
+            {
+                "valid": True,
+                "exists": None,
+                "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
+                "purl": "pkg:rubygems/bundler-sass",
+            },
+            {
+                "valid": True,
+                "exists": None,
+                "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
+                "purl": "pkg:nginx/nginx",
+            },
+            {
+                "valid": True,
+                "exists": False,
+                "message": "The provided PackageURL is valid, but does not exist in the upstream repo.",
+                "purl": "pkg:pypi/zzzzz",
+            },
+            {
+                "valid": False,
+                "exists": None,
+                "message": "The provided PackageURL is not valid.",
+                "purl": "pkg:pypi/?fetchcode",
+            },
+            {
+                "valid": False,
+                "exists": None,
+                "message": "The provided PackageURL is not valid.",
+                "purl": "zzzzz",
+            },
+            {
+                "valid": True,
+                "exists": True,
+                "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+                "purl": "pkg:maven/axis/axis@1.0",
+            },
+        ]
+
+        input_purls_and_expected_states = [
+            ["pkg:pypi/fetchcode", None],
+            ["pkg:gem/bundler-sass", "valid_but_not_supported"],
+            ["pkg:rubygems/bundler-sass", "check_existence_not_supported"],
+            ["pkg:nginx/nginx", "valid_but_not_supported"],
+            ["pkg:pypi/zzzzz", "not_in_upstream_repo"],
+            ["pkg:pypi/?fetchcode", "not_valid"],
+            ["zzzzz", "not_valid"],
+            ["pkg:maven/axis/axis@1.0", "valid_but_not_supported"],
+        ]
+
+        for input_purl, expected_state in input_purls_and_expected_states:
+            purl_metadata = purlcli.check_metadata_purl(input_purl)
+            assert purl_metadata == expected_state
 
     @pytest.mark.parametrize(
         "test_input,expected_input_purls,expected_normalized_purls",
@@ -849,7 +643,7 @@ class TestPURLCLI_metadata(object):
                             "pkg:pypi/fetchcode@0.2.0",
                         ],
                         "tool_name": "purlcli",
-                        "tool_version": "0.2.0",
+                        "tool_version": "0.1.0",
                         "warnings": [
                             "'pkg:gem/bundler-sass' not supported with `metadata` command"
                         ],
@@ -906,7 +700,7 @@ class TestPURLCLI_metadata(object):
                             "pkg:pypi/fetchcode@0.2.0",
                         ],
                         "tool_name": "purlcli",
-                        "tool_version": "0.2.0",
+                        "tool_version": "0.1.0",
                         "warnings": [
                             "input PURL: 'pkg:pypi/fetchcode@0.1.0' normalized to 'pkg:pypi/fetchcode'",
                             "input PURL: 'pkg:pypi/fetchcode@0.2.0' normalized to 'pkg:pypi/fetchcode'",
@@ -940,248 +734,32 @@ class TestPURLCLI_metadata(object):
 
 
 class TestPURLCLI_urls(object):
-    def test_urls_cli(self):
+    @mock.patch("purldb_toolkit.purlcli.make_head_request")
+    def test_urls_cli_head(self, mock_make_head_request):
         """
         Test the `urls` command with actual and expected JSON output files.
         """
+        mock_make_head_request.side_effect = [
+            {"get_request": "N/A"},
+            {"head_request": "N/A"},
+            {"get_request": "N/A"},
+            {"head_request": "N/A"},
+            {"get_request": "N/A"},
+            {"head_request": "N/A"},
+            {"get_request": 200},
+            {"head_request": 200},
+            {"get_request": 200},
+            {"head_request": 200},
+            {"get_request": 200},
+            {"head_request": 200},
+        ]
         expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_urls_output.json"
+            "purlcli/expected_urls_output_head_mock.json"
         )
-        actual_result_file = test_env.get_temp_file("actual_urls_output.json")
+        actual_result_file = test_env.get_temp_file("actual_urls_output_head_mock.json")
         options = [
             "--purl",
             "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:pypi/dejacode",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0?os=windows",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0os=windows",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0?how_is_the_weather=rainy",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0#how/are/you",
-            "--purl",
-            "pkg:pypi/dejacode@10.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:cargo/socksprox",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:nginx/nginx@0.8.9?os=windows",
-            "--purl",
-            "pkg:gem/bundler-sass",
-            "--purl",
-            "pkg:rubygems/bundler-sass",
-            "--purl",
-            "pkg:pypi/matchcode",
-            "--purl",
-            "abcdefg",
-            "--purl",
-            "pkg/abc",
-            "--purl",
-            "pkg:nuget/auth0-aspnet@1.1.0",
-            "--output",
-            actual_result_file,
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.get_urls, options, catch_exceptions=False)
-        assert result.exit_code == 0
-
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-        cli_test_utils.streamline_headers(output_data["headers"])
-
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-        cli_test_utils.streamline_headers(expected_data["headers"])
-
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
-            (output_data["packages"], expected_data["packages"]),
-        ]
-
-        for output, expected in result_objects:
-            assert output == expected
-
-    def test_urls_cli_unique(self):
-        """
-        Test the `urls` command with actual and expected JSON output files.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_urls_output_unique.json"
-        )
-        actual_result_file = test_env.get_temp_file("actual_urls_output_unique.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:pypi/dejacode",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0?os=windows",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0os=windows",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0?how_is_the_weather=rainy",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0#how/are/you",
-            "--purl",
-            "pkg:pypi/dejacode@10.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:cargo/socksprox",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:nginx/nginx@0.8.9?os=windows",
-            "--purl",
-            "pkg:gem/bundler-sass",
-            "--purl",
-            "pkg:rubygems/bundler-sass",
-            "--purl",
-            "pkg:pypi/matchcode",
-            "--purl",
-            "abcdefg",
-            "--purl",
-            "pkg/abc",
-            "--purl",
-            "pkg:nuget/auth0-aspnet@1.1.0",
-            "--output",
-            actual_result_file,
-            "--unique",
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.get_urls, options, catch_exceptions=False)
-        assert result.exit_code == 0
-
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-        cli_test_utils.streamline_headers(output_data["headers"])
-
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-        cli_test_utils.streamline_headers(expected_data["headers"])
-
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
-            (output_data["packages"], expected_data["packages"]),
-        ]
-
-        for output, expected in result_objects:
-            assert output == expected
-
-    def test_urls_cli_head(self):
-        """
-        Test the `urls` command with actual and expected JSON output files.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_urls_output_head.json"
-        )
-        actual_result_file = test_env.get_temp_file("actual_urls_output_head.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:pypi/dejacode",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0?os=windows",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0os=windows",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0?how_is_the_weather=rainy",
-            "--purl",
-            "pkg:pypi/dejacode@5.0.0#how/are/you",
-            "--purl",
-            "pkg:pypi/dejacode@10.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:cargo/socksprox",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:nginx/nginx@0.8.9?os=windows",
-            "--purl",
-            "pkg:gem/bundler-sass",
-            "--purl",
-            "pkg:rubygems/bundler-sass",
-            "--purl",
-            "pkg:pypi/matchcode",
-            "--purl",
-            "abcdefg",
-            "--purl",
-            "pkg/abc",
-            "--purl",
-            "pkg:nuget/auth0-aspnet@1.1.0",
             "--head",
             "--output",
             actual_result_file,
@@ -1190,13 +768,13 @@ class TestPURLCLI_urls(object):
         result = runner.invoke(purlcli.get_urls, options, catch_exceptions=False)
         assert result.exit_code == 0
 
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-        cli_test_utils.streamline_headers(output_data["headers"])
+        with open(actual_result_file) as f_output:
+            output_data = json.load(f_output)
+            cli_test_utils.streamline_headers(output_data["headers"])
 
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-        cli_test_utils.streamline_headers(expected_data["headers"])
+        with open(expected_result_file) as f_expected:
+            expected_data = json.load(f_expected)
+            cli_test_utils.streamline_headers(expected_data["headers"])
 
         result_objects = [
             (
@@ -1264,827 +842,230 @@ class TestPURLCLI_urls(object):
         assert "Use either purls or file." in result.output
         assert result.exit_code == 2
 
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                ["pkg:pypi/fetchcode"],
+    @mock.patch("purldb_toolkit.purlcli.check_urls_purl")
+    def test_urls_details(self, mock_check_urls_purl):
+        mock_check_urls_purl.return_value = "valid_but_not_fully_supported"
+
+        expected_data = {
+            "headers": [
                 {
-                    "headers": [
+                    "tool_name": "purlcli",
+                    "tool_version": "0.1.0",
+                    "options": {
+                        "command": "urls",
+                        "--purl": ["pkg:pypi/fetchcode"],
+                        "--file": None,
+                        "--output": "",
+                    },
+                    "purls": ["pkg:pypi/fetchcode"],
+                    "errors": [],
+                    "warnings": [
+                        "'pkg:pypi/fetchcode' not fully supported with `urls` command"
+                    ],
+                }
+            ],
+            "packages": [
+                {
+                    "purl": "pkg:pypi/fetchcode",
+                    "download_url": {
+                        "url": None,
+                    },
+                    "inferred_urls": [
                         {
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "options": {
-                                "command": "urls",
-                                "--purl": ["pkg:pypi/fetchcode"],
-                                "--file": None,
-                                "--output": "",
-                            },
-                            "purls": ["pkg:pypi/fetchcode"],
-                            "errors": [],
-                            "warnings": [
-                                "'pkg:pypi/fetchcode' not fully supported with `urls` command"
-                            ],
+                            "url": "https://pypi.org/project/fetchcode/",
                         }
                     ],
-                    "packages": [
-                        {
-                            "purl": "pkg:pypi/fetchcode",
-                            "download_url": {
-                                "url": None,
-                            },
-                            "inferred_urls": [
-                                {
-                                    "url": "https://pypi.org/project/fetchcode/",
-                                }
-                            ],
-                            "repo_download_url": {
-                                "url": None,
-                            },
-                            "repo_download_url_by_package_type": {
-                                "url": None,
-                            },
-                            "repo_url": {
-                                "url": "https://pypi.org/project/fetchcode/",
-                            },
-                            "url": {
-                                "url": "https://pypi.org/project/fetchcode/",
-                            },
-                        },
-                    ],
+                    "repo_download_url": {
+                        "url": None,
+                    },
+                    "repo_download_url_by_package_type": {
+                        "url": None,
+                    },
+                    "repo_url": {
+                        "url": "https://pypi.org/project/fetchcode/",
+                    },
+                    "url": {
+                        "url": "https://pypi.org/project/fetchcode/",
+                    },
                 },
-            ),
-            (
-                ["pkg:pypi/fetchcode@10.0.0"],
-                {
-                    "headers": [
-                        {
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "options": {
-                                "command": "urls",
-                                "--purl": ["pkg:pypi/fetchcode@10.0.0"],
-                                "--file": None,
-                                "--output": "",
-                            },
-                            "purls": ["pkg:pypi/fetchcode@10.0.0"],
-                            "errors": [],
-                            "warnings": [
-                                "'pkg:pypi/fetchcode@10.0.0' does not exist in the upstream repo",
-                            ],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["pkg:gem/bundler-sass"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:gem/bundler-sass"],
-                                "command": "urls",
-                            },
-                            "purls": ["pkg:gem/bundler-sass"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": [],
-                        }
-                    ],
-                    "packages": [
-                        {
-                            "purl": "pkg:gem/bundler-sass",
-                            "download_url": {
-                                "url": None,
-                            },
-                            "inferred_urls": [
-                                {
-                                    "url": "https://rubygems.org/gems/bundler-sass",
-                                }
-                            ],
-                            "repo_download_url": {
-                                "url": None,
-                            },
-                            "repo_download_url_by_package_type": {
-                                "url": None,
-                            },
-                            "repo_url": {
-                                "url": "https://rubygems.org/gems/bundler-sass",
-                            },
-                            "url": {
-                                "url": "https://rubygems.org/gems/bundler-sass",
-                            },
-                        },
-                    ],
-                },
-            ),
-            (
-                ["pkg:rubygems/bundler-sass"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:rubygems/bundler-sass"],
-                                "command": "urls",
-                            },
-                            "purls": ["pkg:rubygems/bundler-sass"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": [],
-                        }
-                    ],
-                    "packages": [
-                        {
-                            "purl": "pkg:rubygems/bundler-sass",
-                            "download_url": {
-                                "url": None,
-                            },
-                            "inferred_urls": [
-                                {
-                                    "url": "https://rubygems.org/gems/bundler-sass",
-                                }
-                            ],
-                            "repo_download_url": {
-                                "url": None,
-                            },
-                            "repo_download_url_by_package_type": {
-                                "url": None,
-                            },
-                            "repo_url": {
-                                "url": "https://rubygems.org/gems/bundler-sass",
-                            },
-                            "url": {
-                                "url": "https://rubygems.org/gems/bundler-sass",
-                            },
-                        },
-                    ],
-                },
-            ),
-            (
-                ["pkg:nginx/nginx"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:nginx/nginx"],
-                                "command": "urls",
-                            },
-                            "purls": ["pkg:nginx/nginx"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": [
-                                "'pkg:nginx/nginx' not supported with `urls` command"
-                            ],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["pkg:pypi/zzzzz"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:pypi/zzzzz"],
-                                "command": "urls",
-                            },
-                            "purls": ["pkg:pypi/zzzzz"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": [
-                                "'pkg:pypi/zzzzz' does not exist in the upstream repo",
-                            ],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["pkg:pypi/?fetchcode"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["pkg:pypi/?fetchcode"],
-                                "command": "urls",
-                            },
-                            "purls": ["pkg:pypi/?fetchcode"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": ["'pkg:pypi/?fetchcode' not valid"],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-            (
-                ["zzzzz"],
-                {
-                    "headers": [
-                        {
-                            "errors": [],
-                            "options": {
-                                "--file": None,
-                                "--output": "",
-                                "--purl": ["zzzzz"],
-                                "command": "urls",
-                            },
-                            "purls": ["zzzzz"],
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "warnings": ["'zzzzz' not valid"],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-        ],
-    )
-    def test_urls_details(self, test_input, expected):
-        """
-        Test the `urls` nested function, `get_urls_details()`.
-        """
+            ],
+        }
+
+        input_purls = ["pkg:pypi/fetchcode"]
+
         purl_urls = purlcli.get_urls_details(
-            test_input,
+            input_purls,
             output="",
             file="",
             command_name="urls",
             head=False,
             unique=False,
         )
-        cli_test_utils.streamline_headers(expected["headers"])
+        cli_test_utils.streamline_headers(expected_data["headers"])
         cli_test_utils.streamline_headers(purl_urls["headers"])
 
+        assert purl_urls == expected_data
+
+    @mock.patch("purldb_toolkit.purlcli.validate_purl")
+    def test_check_urls_purl(self, mock_validate_purl):
+        mock_validate_purl.return_value = {
+            "valid": True,
+            "exists": True,
+            "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+            "purl": "pkg:pypi/fetchcode",
+        }
+
+        input_purl = "pkg:pypi/fetchcode"
+        expected = "valid_but_not_fully_supported"
+        purl_urls = purlcli.check_urls_purl(input_purl)
+
         assert purl_urls == expected
 
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                ["pkg:pypi/fetchcode"],
-                "valid_but_not_fully_supported",
-            ),
-            (
-                ["pkg:gem/bundler-sass"],
-                None,
-            ),
-            (
-                ["pkg:rubygems/bundler-sass"],
-                None,
-            ),
-            (
-                ["pkg:nginx/nginx"],
-                "valid_but_not_supported",
-            ),
-            (
-                ["pkg:pypi/zzzzz"],
-                "not_in_upstream_repo",
-            ),
-            (
-                ["pkg:pypi/?fetchcode"],
-                "not_valid",
-            ),
-            (
-                ["zzzzz"],
-                "not_valid",
-            ),
-        ],
-    )
-    def test_check_urls_purl(self, test_input, expected):
-        purl_urls = purlcli.check_urls_purl(test_input[0])
-        assert purl_urls == expected
+    @mock.patch("requests.get")
+    @mock.patch("requests.head")
+    def test_validate_purl_mock_requests_get_and_head(
+        self, mock_requests_head, mock_requests_get
+    ):
+        get_response = mock.Mock(requests.Response)
+        get_response.status_code = 200
 
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                ["https://pypi.org/project/fetchcode/"],
-                {"get_request": 200, "head_request": 200},
-            ),
-            (
-                [None],
-                {"get_request": "N/A", "head_request": "N/A"},
-            ),
-            (
-                ["https://crates.io/crates/banquo"],
-                {"get_request": 404, "head_request": 404},
-            ),
-            (
-                ["https://crates.io/crates/socksprox"],
-                {"get_request": 404, "head_request": 404},
-            ),
-            (
-                ["https://www.nuget.org/api/v2/package/auth0-aspnet/1.1.0"],
-                {"get_request": 200, "head_request": 404},
-            ),
-        ],
-    )
-    def test_make_head_request(self, test_input, expected):
-        purl_status_code = purlcli.make_head_request(test_input[0])
+        head_response = mock.Mock(requests.Response)
+        head_response.status_code = 400
 
-        assert purl_status_code == expected
+        mock_requests_get.return_value = get_response
+        mock_requests_head.return_value = head_response
+
+        expected_results = {"get_request": 200, "head_request": 400}
+
+        url_detail = "https://pypi.org/project/fetchcode/"
+        results = purlcli.make_head_request(url_detail)
+
+        assert results == expected_results
 
 
 class TestPURLCLI_validate(object):
-    def test_validate_cli(self):
-        """
-        Test the `validate` command with actual and expected JSON output files.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_validate_output.json"
+    @mock.patch("requests.get")
+    def test_validate_purl_mock_requests_get(self, mock_requests_get):
+        mock_request_response = {
+            "valid": True,
+            "exists": True,
+            "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+            "purl": "pkg:pypi/fetchcode",
+        }
+
+        def mock_requests_get_return_func():
+            return mock_request_response
+
+        mock_requests_get.return_value.json = mock_requests_get_return_func
+        input_purl = "pkg:pypi/fetchcode"
+        results = purlcli.validate_purl(input_purl)
+        assert mock_request_response == results
+
+    @mock.patch("requests.get")
+    def test_validate_purl_mock_requests_get_jsondecodeerror(self, mock_requests_get):
+        def json_decode_failure_exception(**kwargs):
+            raise json.decoder.JSONDecodeError("test", "[{}]", 0)
+
+        mock_requests_get.return_value.json = json_decode_failure_exception
+        input_purl = "pkg:pypi/fetchcode"
+        out = StringIO()
+        with mock.patch("requests.Response.json", json_decode_failure_exception):
+            with redirect_stdout(out):
+                purlcli.validate_purl(input_purl)
+        results = out.getvalue()
+        assert (
+            "json.decoder.JSONDecodeError for 'pkg:pypi/fetchcode': test: line 1 column 1 (char 0)"
+            in results
         )
-        actual_result_file = test_env.get_temp_file("actual_validate_output.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0?os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:gem/rails",
-            "--purl",
-            "pkg:rubygems/rails",
-            "--output",
-            actual_result_file,
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.validate, options, catch_exceptions=False)
-        assert result.exit_code == 0
 
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
+    @mock.patch("requests.get")
+    def test_validate_purl_mock_requests_get_exception(self, mock_requests_get):
+        def raise_exception(**kwargs):
+            raise Exception
 
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
+        mock_requests_get.return_value.json = raise_exception
+        input_purl = "pkg:pypi/fetchcode"
+        out = StringIO()
+        with mock.patch("requests.Response.json", raise_exception):
+            with redirect_stdout(out):
+                purlcli.validate_purl(input_purl)
+        results = out.getvalue()
+        assert "'validate' endpoint error for 'pkg:pypi/fetchcode': \n" in results
 
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
-            (output_data["packages"], expected_data["packages"]),
-        ]
+    @mock.patch("requests.get")
+    def test_validate_purl_mock_requests_none(self, mock_requests_get):
+        def raise_exception(**kwargs):
+            raise Exception
 
-        for output, expected in result_objects:
-            assert output == expected
-
-    def test_validate_cli_unique(self):
-        """
-        Test the `validate` command with actual and expected JSON output files
-        with the `--unique` flag included in the command.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_validate_output_unique.json"
-        )
-        actual_result_file = test_env.get_temp_file("actual_validate_output.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0?os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:gem/rails",
-            "--purl",
-            "pkg:rubygems/rails",
-            "--output",
-            actual_result_file,
-            "--unique",
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.validate, options, catch_exceptions=False)
-        assert result.exit_code == 0
-
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--unique"],
-                expected_data["headers"][0]["options"]["--unique"],
-            ),
-            (output_data["packages"], expected_data["packages"]),
-        ]
-
-        for output, expected in result_objects:
-            assert output == expected
-
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                "pkg:pypi/fetchcode@0.2.0",
-                {
-                    "valid": True,
-                    "exists": True,
-                    "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
-                    "purl": "pkg:pypi/fetchcode@0.2.0",
-                },
-            ),
-            (
-                "pkg:pypi/fetchcode@10.2.0",
-                {
-                    "valid": True,
-                    "exists": False,
-                    "message": "The provided PackageURL is valid, but does not exist in the upstream repo.",
-                    "purl": "pkg:pypi/fetchcode@10.2.0",
-                },
-            ),
-            (
-                "pkg:nginx/nginx@0.8.9?os=windows",
-                {
-                    "valid": True,
-                    "exists": None,
-                    "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
-                    "purl": "pkg:nginx/nginx@0.8.9?os=windows",
-                },
-            ),
-            (
-                "pkg:gem/bundler-sass",
-                {
-                    "valid": True,
-                    "exists": True,
-                    "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
-                    "purl": "pkg:gem/bundler-sass",
-                },
-            ),
-            (
-                "pkg:rubygems/bundler-sass",
-                {
-                    "valid": True,
-                    "exists": None,
-                    "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
-                    "purl": "pkg:rubygems/bundler-sass",
-                },
-            ),
-            (
-                "pkg:maven/com.fasterxml.jackson.core/jackson-databind@2.14.0-rc1",
-                {
-                    "valid": True,
-                    "exists": True,
-                    "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
-                    "purl": "pkg:maven/com.fasterxml.jackson.core/jackson-databind@2.14.0-rc1",
-                },
-            ),
-        ],
-    )
-    def test_validate_purl(self, test_input, expected):
-        validated_purl = purlcli.validate_purl(test_input)
-        assert validated_purl == expected
-
-    def test_validate_purl_empty(self):
-        test_input = None
-        validated_purl = purlcli.validate_purl(test_input)
-        expected_results = {"errors": {"purl": ["This field is required."]}}
-        assert validated_purl == expected_results
-
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                "pkg:pypi/fetchcode@0.2.0",
-                {
-                    "valid": True,
-                    "exists": True,
-                    "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
-                    "purl": "pkg:pypi/fetchcode@0.2.0",
-                },
-            ),
-            (
-                "pkg:pypi/fetchcode@0.2.0?",
-                {
-                    "valid": True,
-                    "exists": True,
-                    "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
-                    "purl": "pkg:pypi/fetchcode@0.2.0?",
-                },
-            ),
-            (
-                "pkg:pypi/fetchcode@?0.2.0",
-                {
-                    "valid": False,
-                    "exists": None,
-                    "message": "The provided PackageURL is not valid.",
-                    "purl": "pkg:pypi/fetchcode@?0.2.0",
-                },
-            ),
-            (
-                "foo",
-                {
-                    "valid": False,
-                    "exists": None,
-                    "message": "The provided PackageURL is not valid.",
-                    "purl": "foo",
-                },
-            ),
-        ],
-    )
-    def test_validate_purl_invalid(self, test_input, expected):
-        validated_purl = purlcli.validate_purl(test_input)
-        assert validated_purl == expected
-
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                "pkg:nginx/nginx@0.8.9?os=windows",
-                {
-                    "valid": True,
-                    "exists": None,
-                    "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
-                    "purl": "pkg:nginx/nginx@0.8.9?os=windows",
-                },
-            ),
-            (
-                " pkg:nginx/nginx@0.8.9?os=windows",
-                {
-                    "valid": True,
-                    "exists": None,
-                    "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
-                    "purl": "pkg:nginx/nginx@0.8.9?os=windows",
-                },
-            ),
-            (
-                "pkg:nginx/nginx@0.8.9?os=windows ",
-                {
-                    "valid": True,
-                    "exists": None,
-                    "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
-                    "purl": "pkg:nginx/nginx@0.8.9?os=windows",
-                },
-            ),
-        ],
-    )
-    def test_validate_purl_strip(self, test_input, expected):
-        validated_purl = purlcli.validate_purl(test_input)
-        assert validated_purl == expected
+        mock_requests_get.return_value.json = raise_exception
+        input_purl = None
+        out = StringIO()
+        with mock.patch("requests.Response.json", raise_exception):
+            with redirect_stdout(out):
+                purlcli.validate_purl(input_purl)
+        results = out.getvalue()
+        assert "'validate' endpoint error for 'None': \n" in results
 
 
 class TestPURLCLI_versions(object):
-    def test_versions_cli(self):
-        """
-        Test the `versions` command with actual and expected JSON output files.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_versions_output.json"
-        )
-        actual_result_file = test_env.get_temp_file("actual_versions_output.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0?os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:hex/coherence@0.1.0",
-            "--output",
-            actual_result_file,
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.get_versions, options, catch_exceptions=False)
-        assert result.exit_code == 0
+    @mock.patch("purldb_toolkit.purlcli.collect_versions")
+    @mock.patch("purldb_toolkit.purlcli.check_versions_purl")
+    def test_versions_details_multiple(
+        self,
+        mock_check_versions_purl,
+        mock_collect_versions,
+    ):
 
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-        cli_test_utils.streamline_headers(output_data["headers"])
-
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-        cli_test_utils.streamline_headers(expected_data["headers"])
-
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
+        mock_check_versions_purl.side_effect = [
+            None,
+            None,
+            "valid_but_not_supported",
+            "valid_but_not_supported",
+            None,
+            "not_valid",
         ]
 
-        for output, expected in result_objects:
-            assert output == expected
-
-        expected_versions_01 = []
-        for pkg in expected_data["packages"]:
-            expected_versions_01.append(pkg)
-
-        actual_versions_01 = []
-        for pkg in output_data["packages"]:
-            actual_versions_01.append(pkg)
-
-        for exp in expected_versions_01:
-            for actual in actual_versions_01:
-                if exp["purl"] == actual["purl"]:
-                    for exp_vers in exp["versions"]:
-                        assert exp_vers in actual["versions"]
-
-    def test_versions_cli_unique(self):
-        """
-        Test the `versions` command with actual and expected JSON output files
-        with the `--unique` flag included in the command.
-        """
-        expected_result_file = test_env.get_test_loc(
-            "purlcli/expected_versions_output_unique.json"
-        )
-        actual_result_file = test_env.get_temp_file("actual_versions_output.json")
-        options = [
-            "--purl",
-            "pkg:pypi/fetchcode",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0?os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@0.3.0os=windows",
-            "--purl",
-            "pkg:pypi/fetchcode@5.0.0",
-            "--purl",
-            "pkg:cargo/banquo",
-            "--purl",
-            "pkg:nginx/nginx",
-            "--purl",
-            "pkg:hex/coherence@0.1.0",
-            "--output",
-            actual_result_file,
-            "--unique",
-        ]
-        runner = CliRunner()
-        result = runner.invoke(purlcli.get_versions, options, catch_exceptions=False)
-        assert result.exit_code == 0
-
-        f_output = open(actual_result_file)
-        output_data = json.load(f_output)
-        cli_test_utils.streamline_headers(output_data["headers"])
-
-        f_expected = open(expected_result_file)
-        expected_data = json.load(f_expected)
-        cli_test_utils.streamline_headers(expected_data["headers"])
-
-        result_objects = [
-            (
-                output_data["headers"][0]["tool_name"],
-                expected_data["headers"][0]["tool_name"],
-            ),
-            (output_data["headers"][0]["purls"], expected_data["headers"][0]["purls"]),
-            (
-                output_data["headers"][0]["warnings"],
-                expected_data["headers"][0]["warnings"],
-            ),
-            (
-                output_data["headers"][0]["errors"],
-                expected_data["headers"][0]["errors"],
-            ),
-            (
-                output_data["headers"][0]["options"]["command"],
-                expected_data["headers"][0]["options"]["command"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--purl"],
-                expected_data["headers"][0]["options"]["--purl"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--file"],
-                expected_data["headers"][0]["options"]["--file"],
-            ),
-            (
-                output_data["headers"][0]["options"]["--unique"],
-                expected_data["headers"][0]["options"]["--unique"],
-            ),
+        mock_collect_versions.side_effect = [
+            [
+                {
+                    "purl": "pkg:pypi/fetchcode@0.1.0",
+                    "version": "0.1.0",
+                    "release_date": "2021-08-25 15:15:15.265015+00:00",
+                },
+                {
+                    "purl": "pkg:pypi/fetchcode@0.2.0",
+                    "version": "0.2.0",
+                    "release_date": "2022-09-14 16:36:02.242182+00:00",
+                },
+                {
+                    "purl": "pkg:pypi/fetchcode@0.3.0",
+                    "version": "0.3.0",
+                    "release_date": "2023-12-18 20:49:45.840364+00:00",
+                },
+            ],
+            [
+                {
+                    "purl": "pkg:gem/bundler-sass@0.1.2",
+                    "version": "0.1.2",
+                    "release_date": "2013-12-11 00:27:10.097000+00:00",
+                }
+            ],
+            [
+                {
+                    "purl": "pkg:cargo/socksprox@0.1.1",
+                    "release_date": "2024-02-07 23:29:58.801293+00:00",
+                    "version": "0.1.1",
+                },
+                {
+                    "purl": "pkg:cargo/socksprox@0.1.0",
+                    "release_date": "2024-02-07 23:21:05.242366+00:00",
+                    "version": "0.1.0",
+                },
+            ],
         ]
 
-        for output, expected in result_objects:
-            assert output == expected
-
-        expected_versions_01 = []
-        for pkg in expected_data["packages"]:
-            expected_versions_01.append(pkg)
-
-        actual_versions_01 = []
-        for pkg in output_data["packages"]:
-            actual_versions_01.append(pkg)
-
-        for exp in expected_versions_01:
-            for actual in actual_versions_01:
-                if exp["purl"] == actual["purl"]:
-                    for exp_vers in exp["versions"]:
-                        assert exp_vers in actual["versions"]
-
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
+        input_purls_and_expected_purl_data = [
+            [
                 ["pkg:pypi/fetchcode"],
                 {
                     "headers": [
@@ -2121,17 +1102,12 @@ class TestPURLCLI_versions(object):
                                     "version": "0.3.0",
                                     "release_date": "2023-12-18 20:49:45.840364+00:00",
                                 },
-                                {
-                                    "purl": "pkg:pypi/fetchcode@0.4.0",
-                                    "release_date": "2024-03-12 07:01:29.239299+00:00",
-                                    "version": "0.4.0",
-                                },
                             ],
                         }
                     ],
                 },
-            ),
-            (
+            ],
+            [
                 ["pkg:gem/bundler-sass"],
                 {
                     "headers": [
@@ -2162,8 +1138,8 @@ class TestPURLCLI_versions(object):
                         }
                     ],
                 },
-            ),
-            (
+            ],
+            [
                 ["pkg:rubygems/bundler-sass"],
                 {
                     "headers": [
@@ -2185,8 +1161,8 @@ class TestPURLCLI_versions(object):
                     ],
                     "packages": [],
                 },
-            ),
-            (
+            ],
+            [
                 ["pkg:nginx/nginx"],
                 {
                     "headers": [
@@ -2208,9 +1184,9 @@ class TestPURLCLI_versions(object):
                     ],
                     "packages": [],
                 },
-            ),
-            (
-                ["pkg:pypi/zzzzz"],
+            ],
+            [
+                ["pkg:cargo/socksprox"],
                 {
                     "headers": [
                         {
@@ -2218,21 +1194,35 @@ class TestPURLCLI_versions(object):
                             "tool_version": "0.2.0",
                             "options": {
                                 "command": "versions",
-                                "--purl": ["pkg:pypi/zzzzz"],
+                                "--purl": ["pkg:cargo/socksprox"],
                                 "--file": None,
                                 "--output": "",
                             },
-                            "purls": ["pkg:pypi/zzzzz"],
+                            "purls": ["pkg:cargo/socksprox"],
                             "errors": [],
-                            "warnings": [
-                                "'pkg:pypi/zzzzz' does not exist in the upstream repo"
+                            "warnings": [],
+                        }
+                    ],
+                    "packages": [
+                        {
+                            "purl": "pkg:cargo/socksprox",
+                            "versions": [
+                                {
+                                    "purl": "pkg:cargo/socksprox@0.1.1",
+                                    "version": "0.1.1",
+                                    "release_date": "2024-02-07 23:29:58.801293+00:00",
+                                },
+                                {
+                                    "purl": "pkg:cargo/socksprox@0.1.0",
+                                    "version": "0.1.0",
+                                    "release_date": "2024-02-07 23:21:05.242366+00:00",
+                                },
                             ],
                         }
                     ],
-                    "packages": [],
                 },
-            ),
-            (
+            ],
+            [
                 ["pkg:pypi/?fetchcode"],
                 {
                     "headers": [
@@ -2252,98 +1242,182 @@ class TestPURLCLI_versions(object):
                     ],
                     "packages": [],
                 },
-            ),
-            (
-                ["zzzzz"],
-                {
-                    "headers": [
-                        {
-                            "tool_name": "purlcli",
-                            "tool_version": "0.2.0",
-                            "options": {
-                                "command": "versions",
-                                "--purl": ["zzzzz"],
-                                "--file": None,
-                                "--output": "",
-                            },
-                            "purls": ["zzzzz"],
-                            "errors": [],
-                            "warnings": ["'zzzzz' not valid"],
-                        }
-                    ],
-                    "packages": [],
-                },
-            ),
-        ],
-    )
-    def test_versions_details(self, test_input, expected):
+            ],
+        ]
+
         output = ""
         file = ""
         command_name = "versions"
         unique = False
 
-        purl_versions = purlcli.get_versions_details(
-            test_input, output, file, unique, command_name
+        for input_purl, expected_data in input_purls_and_expected_purl_data:
+            purl_versions_data = purlcli.get_versions_details(
+                input_purl, output, file, unique, command_name
+            )
+
+            assert purl_versions_data == expected_data
+
+    @mock.patch("purldb_toolkit.purlcli.collect_versions")
+    @mock.patch("purldb_toolkit.purlcli.check_versions_purl")
+    def test_versions_details(
+        self,
+        mock_check_versions_purl,
+        mock_collect_versions,
+    ):
+
+        mock_collect_versions.return_value = [
+            {
+                "purl": "pkg:pypi/fetchcode@0.1.0",
+                "version": "0.1.0",
+                "release_date": "2021-08-25 15:15:15.265015+00:00",
+            },
+            {
+                "purl": "pkg:pypi/fetchcode@0.2.0",
+                "version": "0.2.0",
+                "release_date": "2022-09-14 16:36:02.242182+00:00",
+            },
+            {
+                "purl": "pkg:pypi/fetchcode@0.3.0",
+                "version": "0.3.0",
+                "release_date": "2023-12-18 20:49:45.840364+00:00",
+            },
+        ]
+
+        mock_check_versions_purl.return_value = None
+
+        expected_data = {
+            "headers": [
+                {
+                    "tool_name": "purlcli",
+                    "tool_version": "0.2.0",
+                    "options": {
+                        "command": "versions",
+                        "--purl": ["pkg:pypi/fetchcode"],
+                        "--file": None,
+                        "--output": "",
+                    },
+                    "purls": ["pkg:pypi/fetchcode"],
+                    "errors": [],
+                    "warnings": [],
+                }
+            ],
+            "packages": [
+                {
+                    "purl": "pkg:pypi/fetchcode",
+                    "versions": [
+                        {
+                            "purl": "pkg:pypi/fetchcode@0.1.0",
+                            "version": "0.1.0",
+                            "release_date": "2021-08-25 15:15:15.265015+00:00",
+                        },
+                        {
+                            "purl": "pkg:pypi/fetchcode@0.2.0",
+                            "version": "0.2.0",
+                            "release_date": "2022-09-14 16:36:02.242182+00:00",
+                        },
+                        {
+                            "purl": "pkg:pypi/fetchcode@0.3.0",
+                            "version": "0.3.0",
+                            "release_date": "2023-12-18 20:49:45.840364+00:00",
+                        },
+                    ],
+                }
+            ],
+        }
+
+        input_purls = ["pkg:pypi/fetchcode"]
+
+        output = ""
+        file = ""
+        command_name = "versions"
+        unique = False
+
+        purl_versions_data = purlcli.get_versions_details(
+            input_purls,
+            output,
+            file,
+            unique,
+            command_name,
         )
+        assert purl_versions_data == expected_data
 
-        cli_test_utils.streamline_headers(purl_versions["headers"])
-        cli_test_utils.streamline_headers(expected["headers"])
+    @mock.patch("purldb_toolkit.purlcli.validate_purl")
+    def test_check_versions_purl_multiple(self, mock_validate_purl):
+        mock_validate_purl.side_effect = [
+            {
+                "valid": True,
+                "exists": True,
+                "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+                "purl": "pkg:pypi/fetchcode",
+            },
+            {
+                "valid": True,
+                "exists": True,
+                "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+                "purl": "pkg:gem/bundler-sass",
+            },
+            {
+                "valid": True,
+                "exists": None,
+                "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
+                "purl": "pkg:rubygems/bundler-sass",
+            },
+            {
+                "valid": True,
+                "exists": None,
+                "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
+                "purl": "pkg:nginx/nginx",
+            },
+            {
+                "valid": True,
+                "exists": False,
+                "message": "The provided PackageURL is valid, but does not exist in the upstream repo.",
+                "purl": "pkg:pypi/zzzzz",
+            },
+            {
+                "valid": False,
+                "exists": None,
+                "message": "The provided PackageURL is not valid.",
+                "purl": "pkg:pypi/?fetchcode",
+            },
+            {
+                "valid": False,
+                "exists": None,
+                "message": "The provided PackageURL is not valid.",
+                "purl": "zzzzz",
+            },
+            {
+                "valid": True,
+                "exists": True,
+                "message": "The provided Package URL is valid, and the package exists in the upstream repo.",
+                "purl": "pkg:maven/axis/axis@1.0",
+            },
+        ]
+        input_purls_and_expected_states = [
+            ["pkg:pypi/fetchcode", None],
+            ["pkg:gem/bundler-sass", None],
+            ["pkg:rubygems/bundler-sass", "valid_but_not_supported"],
+            ["pkg:nginx/nginx", "valid_but_not_supported"],
+            ["pkg:pypi/zzzzz", "not_in_upstream_repo"],
+            ["pkg:pypi/?fetchcode", "not_valid"],
+            ["zzzzz", "not_valid"],
+            ["pkg:maven/axis/axis@1.0", None],
+        ]
+        for input_purl, expected_state in input_purls_and_expected_states:
+            purl_versions = purlcli.check_versions_purl(input_purl)
+            assert purl_versions == expected_state
 
-        assert purl_versions["headers"] == expected["headers"]
-
-        expected_versions = []
-        for pkg in expected["packages"]:
-            expected_versions.append(pkg)
-
-        actual_versions = []
-        for pkg in purl_versions["packages"]:
-            actual_versions.append(pkg)
-
-        for exp in expected_versions:
-            for actual in actual_versions:
-                if exp["purl"] == actual["purl"]:
-                    for exp_versions in exp["versions"]:
-                        assert exp_versions in actual["versions"]
-
-    @pytest.mark.parametrize(
-        "test_input,expected",
-        [
-            (
-                ["pkg:pypi/fetchcode"],
-                None,
-            ),
-            (
-                ["pkg:gem/bundler-sass"],
-                None,
-            ),
-            (
-                ["pkg:rubygems/bundler-sass"],
-                "valid_but_not_supported",
-            ),
-            (
-                ["pkg:nginx/nginx"],
-                "valid_but_not_supported",
-            ),
-            (
-                ["pkg:pypi/zzzzz"],
-                "not_in_upstream_repo",
-            ),
-            (
-                ["pkg:pypi/?fetchcode"],
-                "not_valid",
-            ),
-            (
-                ["zzzzz"],
-                "not_valid",
-            ),
-            (
-                ["pkg:maven/axis/axis@1.0"],
-                None,
-            ),
-        ],
-    )
-    def test_check_versions_purl(self, test_input, expected):
-        purl_versions = purlcli.check_versions_purl(test_input[0])
+    @mock.patch("purldb_toolkit.purlcli.validate_purl")
+    def test_check_versions_purl(self, mock_validate_purl):
+        mock_validate_purl.return_value = {
+            "valid": True,
+            "exists": None,
+            "message": "The provided PackageURL is valid, but `check_existence` is not supported for this package type.",
+            "purl": "pkg:rubygems/bundler-sass",
+        }
+        input_purl = "pkg:rubygems/bundler-sass"
+        purl_versions = purlcli.check_versions_purl(input_purl)
+        expected = "valid_but_not_supported"
         assert purl_versions == expected
